+++ packages/core/workerfarm/src/Worker.js
@@ -2,10 +2,11 @@ const childProcess = require('child_process');
 const {EventEmitter} = require('events');
 const errorUtils = require('./errorUtils');
 
-const childModule =
+const childModule = require.resolve('./child');
+/*
   parseInt(process.versions.node, 10) < 8
     ? require.resolve('../../lib/workerfarm/child')
-    : require.resolve('../../src/workerfarm/child');
+    : require.resolve('../../src/workerfarm/child');*/
 
 let WORKER_ID = 0;
