+++ packages/core/workerfarm/src/cpuCount.js
