+++ packages/core/workerfarm/test/integration/workerfarm/master-process-id.js
@@ -1,3 +1,3 @@
 module.exports = function() {
   return process.pid;
-}
+};
