+++ packages/core/workerfarm/test/integration/workerfarm/master-sum.js
@@ -1,3 +1,3 @@
 module.exports = function(a, b) {
   return a + b;
-}
+};
